+++ config/thumbnail.php
@@ -1,8 +1,15 @@
 <?php
 
 return [
+    /**
+     * The path relative to storage/app, 
+     * should match with `path` given below,
+     * it is used by Storage facade
+     */
+    "relative_path" => "public/thumbnails",
 
     "path" => storage_path("app/public/thumbnails"),
+
 
     "asset" => "storage/thumbnails",